requirements.txt
Flask>=2.3
Werkzeug>=2.3
Pillow>=10.0
torch>=2.1
torchvision>=0.16
numpy>=1.26
scikit-learn>=1.3
pandas>=2.3

